MAINT: change frame dtype to uint8

+++ testbeam_analysis/converter/pybar_fei4_converter.py
@@ -24,7 +24,7 @@ logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(name)s - [%(leve
 
 testbeam_analysis_dtype = np.dtype([
     ('event_number', np.int64),
-    ('frame', np.uint32),
+    ('frame', np.uint8),
     ('column', np.uint16),
     ('row', np.uint16),
     ('charge', np.uint16)])
